+++ src/Service.php
@@ -296,7 +296,7 @@ class Service extends Component
         }
 
         foreach (array_keys($array) as $key) {
-            if (!property_exists(EmbeddedAsset::class, $key)) {
+            if (!property_exists(EmbeddedAsset::class, $key) && !EmbeddedAsset::propertyIsDeprecated($key)) {
                 return null;
             }
         }